requirements.txt
hostapd
dnsmasq
lighttpd
php7.0-cgi
python-scapy
